requirements.txt
aio-pika
aiomysql
PyJWT
bcrypt